% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla.R
\name{calc_lambda}
\alias{calc_lambda}
\title{Calculate lambda}
\usage{
calc_lambda(cov_f_inv, cov_fb, mu_f, wgt_b, i, b_i)
}
\arguments{
\item{cov_f_inv}{inverse of covaraince of free weights}

\item{cov_fb}{covariance subset of free weight row and bound weight column}

\item{mu_f}{expected returns of free weights}

\item{wgt_b}{weights of bound assets}

\item{i}{position in lambda_i}

\item{b_i}{upper or lowerbound weight depending on intermediate calc}
}
\description{
Calculate lambda
}
