% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla.R
\name{calc_port_mu}
\alias{calc_port_mu}
\title{Portfolio return}
\usage{
calc_port_mu(mu, w)
}
\arguments{
\item{mu}{expected return column matrix}

\item{w}{portfolio weight column}
}
\description{
Portfolio return
}
