% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla.R
\name{calc_port_vol}
\alias{calc_port_vol}
\title{Portfolio volatility}
\usage{
calc_port_vol(cov_mat, w)
}
\arguments{
\item{cov_mat}{expected covariance matrix}

\item{w}{portfolio weight column}
}
\description{
Portfolio volatility
}
