% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla.R
\name{calc_target_vol}
\alias{calc_target_vol}
\title{Solve weights for a target volatility}
\usage{
calc_target_vol(store, cov_mat, target_vol, tol = 1e-04, max_iter = 10000)
}
\arguments{
\item{store}{output of \code{run_cla}}

\item{cov_mat}{covariance matrix}

\item{target_vol}{volatiltiy level to solve for}

\item{tol}{tolerance for volatility solution difference from \code{target_vol}}

\item{max_iter}{maximum iterations for attempted solution}
}
\description{
Solve weights for a target volatility
}
