% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla.R
\name{calc_wgt_f}
\alias{calc_wgt_f}
\title{Calculate free weights}
\usage{
calc_wgt_f(cov_f_inv, cov_fb, mu_f, wgt_b, l)
}
\arguments{
\item{cov_f_inv}{inverse of covaraince of free weights}

\item{cov_fb}{covariance subset of free weight row and bound weight column}

\item{mu_f}{expected returns of free weights}

\item{wgt_b}{weights of bound assets}

\item{l}{lambda}
}
\description{
Calculate free weights
}
