% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla.R
\name{clean_store}
\alias{clean_store}
\title{Remove weight solutions that violate the lowerbound constraint}
\usage{
clean_store(store)
}
\arguments{
\item{store}{storage list from \code{run_cla}}
}
\description{
Remove weight solutions that violate the lowerbound constraint
}
\note{
Poorly conditioned covariance matrices can lead to solutions with negative
weights.
}
