% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla.R
\name{get_b}
\alias{get_b}
\title{Get bound weights from free weights}
\usage{
get_b(f, n_assets)
}
\arguments{
\item{f}{free weight positions}

\item{n_assets}{total number of assets}
}
\description{
Get bound weights from free weights
}
