% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla.R
\name{init_cla}
\alias{init_cla}
\title{Initialize Critical Line Algo}
\usage{
init_cla(mu_vec, low_bound, up_bound)
}
\arguments{
\item{mu_vec}{vector or column matrix of expected returns}

\item{low_bound}{lower bound weights}

\item{up_bound}{upper bound weights}
}
\value{
list containing initial weight vector and free weight position
}
\description{
Initialize Critical Line Algo
}
