% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla.R
\name{sub_mat}
\alias{sub_mat}
\title{Subset matrices into free and bound sections}
\usage{
sub_mat(mu_vec, cov_mat, wgt_vec, f)
}
\arguments{
\item{mu_vec}{expected returns}

\item{cov_mat}{expected covariance matrix}

\item{wgt_vec}{asset weights}

\item{f}{free weight positions}
}
\description{
Subset matrices into free and bound sections
}
