% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cla.R
\name{run_cla}
\alias{run_cla}
\title{Run the critical line algo}
\usage{
run_cla(mu_vec, cov_mat, low_bound = 0, up_bound = 1)
}
\arguments{
\item{mu_vec}{vector or column matrix of expected returns}

\item{cov_mat}{expected covariance matrix}

\item{low_bound}{lower bound weights, see details}

\item{up_bound}{upper bound weights, see details}
}
\value{
a list containing the weights for each corner portfolio \code{wgt_list}
}
\description{
Run the critical line algo
}
\details{
The \code{low_bound} and \code{up_bound} weights are the corresponding 
upper and lower bound weights for each asset in the \code{mu_vec}. They can
be entered as a single numeric value if all assets have the same upper or lower
bound weight or a vector of upper and lower bound weights can be entered.
}
